Small code update
